Protected the training/selection page, only the authenticated student can access the page.
